Final Report/Team10/NestWise-main/backend-langgraph/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-openai
langgraph
tiktoken
numpy
scikit-learn
pypdf
unstructured
pdfminer.six
pydantic[email]
dnspython
python-jose